--- allel/model/ndarray/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, print_function, division
-
-from . import arrays
-from . import indexes
-from . import tables
-from .arrays import GenotypeArray, GenotypeVector, HaplotypeArray, AlleleCountsArray
-from .indexes import SortedIndex, UniqueIndex, SortedMultiIndex
-from .tables import VariantTable, FeatureTable
